refactored logger module
