+++ logger/src/main/java/executor/service/logger/aspect/LoggingAspect.java
@@ -1,4 +1,4 @@
-package executor.service.logger;
+package executor.service.logger.aspect;
 
 import lombok.RequiredArgsConstructor;
 import org.aspectj.lang.JoinPoint;
@@ -15,7 +15,7 @@ import org.springframework.stereotype.Component;
 public class LoggingAspect {
     private final Logger logger;
 
-    @Before("@within(executor.service.annotation.Logged)")
+    @Before("@within(executor.service.logger.annotation.Logged)")
     public void logBefore(JoinPoint joinPoint) {
         String methodName = joinPoint.getSignature().getName();
         String className = joinPoint.getTarget().getClass().getSimpleName();
@@ -23,7 +23,7 @@ public class LoggingAspect {
         logger.info(LogMessage.EXECUTING_METHOD.getMessage(), methodName, className, args);
     }
 
-    @After("@within(executor.service.annotation.Logged)")
+    @After("@within(executor.service.logger.annotation.Logged)")
     public void logAfter(JoinPoint joinPoint) {
         String className = joinPoint.getTarget().getClass().getSimpleName();
         String methodName = joinPoint.getSignature().getName();
