+++ logger/src/main/java/executor/service/logger/config/LoggerConfiguration.java
@@ -1,4 +1,4 @@
-package executor.service.config;
+package executor.service.logger.config;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
